Separated the tests
