+++ sources/packages/backend/apps/api/src/route-controllers/report/_tests_/e2e/report.aest.controller.exportReport.e2e-spec.ts
@@ -52,11 +52,6 @@ import {
   buildUnmetNeedReportData,
   createApplicationsDataSetup,
 } from "./unmet-need-report-utils";
-import {
-  buildApplicationsByInstitutionData,
-  createApplicationsByInstitutionDataSetup,
-} from "./student-applications-by-institution-report-utils";
-import { ConfigService } from "@sims/utilities/config/config.service";
 
 describe("ReportAestController(e2e)-exportReport", () => {
   let app: INestApplication;
@@ -64,7 +59,6 @@ describe("ReportAestController(e2e)-exportReport", () => {
   let appDataSource: DataSource;
   let sharedCASSupplierUpdatedStudent: Student;
   let casSupplierMaintenanceUpdatesPayload: MinistryReportsFilterAPIInDTO;
-  let applicationArchiveDays: number;
 
   beforeAll(async () => {
     const { nestApplication, dataSource } = await createTestingAppModule();
@@ -78,7 +72,6 @@ describe("ReportAestController(e2e)-exportReport", () => {
       reportName: "CAS_Supplier_Maintenance_Updates_Report",
       params: {},
     };
-    applicationArchiveDays = app.get(ConfigService).applicationArchiveDays;
   });
 
   it("Should generate the eCert Feedback Errors report when a report generation request is made with the appropriate offering intensity and date range.", async () => {
@@ -1674,176 +1667,6 @@ describe("ReportAestController(e2e)-exportReport", () => {
         });
     },
   );
-
-  it("Should generate the Student Applications By Institution Report for ministry when a report generation request is made with the appropriate filters, including optional program.", async () => {
-    // Arrange
-    const now = new Date();
-    const student = await saveFakeStudent(db.dataSource);
-    const institution = await db.institution.save(createFakeInstitution());
-    const application = await createApplicationsByInstitutionDataSetup(db, {
-      student,
-      institution,
-      originalSubmissionDate: now,
-    });
-    // Application with a different program is created to ensure that the report filters out
-    // applications using the program, when a program filter is provided.
-    await createApplicationsByInstitutionDataSetup(db, {
-      student,
-      institution,
-      originalSubmissionDate: now,
-    });
-    const payload = {
-      reportName: "Ministry_Student_Applications_By_Institution_Report",
-      params: {
-        institution: institution.id,
-        program: application.currentAssessment.offering.educationProgram.id,
-        startDate: now,
-        endDate: addDays(1, now),
-        isLimitedByArchiveDate: true,
-        offeringIntensity: {
-          "Full Time": true,
-          "Part Time": true,
-        },
-      },
-    };
-
-    const endpoint = "/aest/report";
-    const ministryUserToken = await getAESTToken(
-      AESTGroups.BusinessAdministrators,
-    );
-    // Expected report records.
-    const expectedRecord = buildApplicationsByInstitutionData(application);
-
-    // Act/Assert
-    await request(app.getHttpServer())
-      .post(endpoint)
-      .send(payload)
-      .auth(ministryUserToken, BEARER_AUTH_TYPE)
-      .expect(HttpStatus.CREATED)
-      .then((response) => {
-        const fileContent = response.request.res["text"];
-        const parsedResult = parse(fileContent, {
-          header: true,
-        });
-        expect(parsedResult.data).toEqual(
-          expect.arrayContaining([expectedRecord]),
-        );
-      });
-  });
-
-  it("Should generate the Student Applications By Institution Report for ministry when a report generation request is made with the appropriate filters, without optional program.", async () => {
-    // Arrange
-    const now = new Date();
-    const student = await saveFakeStudent(db.dataSource);
-    const institution = await db.institution.save(createFakeInstitution());
-    const applications = await Promise.all([
-      createApplicationsByInstitutionDataSetup(db, {
-        student,
-        institution,
-        originalSubmissionDate: now,
-      }),
-      createApplicationsByInstitutionDataSetup(db, {
-        student,
-        institution,
-        originalSubmissionDate: now,
-      }),
-    ]);
-    const payload = {
-      reportName: "Ministry_Student_Applications_By_Institution_Report",
-      params: {
-        institution: institution.id,
-        program: "",
-        startDate: now,
-        endDate: addDays(1, now),
-        isLimitedByArchiveDate: true,
-        offeringIntensity: {
-          "Full Time": true,
-          "Part Time": true,
-        },
-      },
-    };
-
-    const endpoint = "/aest/report";
-    const ministryUserToken = await getAESTToken(
-      AESTGroups.BusinessAdministrators,
-    );
-    // Expected report records.
-    const expectedRecords = applications.map((application) =>
-      buildApplicationsByInstitutionData(application),
-    );
-
-    // Act/Assert
-    await request(app.getHttpServer())
-      .post(endpoint)
-      .send(payload)
-      .auth(ministryUserToken, BEARER_AUTH_TYPE)
-      .expect(HttpStatus.CREATED)
-      .then((response) => {
-        const fileContent = response.request.res["text"];
-        const parsedResult = parse(fileContent, {
-          header: true,
-        });
-        expect(parsedResult.data).toEqual(
-          expect.arrayContaining(expectedRecords),
-        );
-      });
-  });
-
-  it("Should generate the Student Applications By Institution Report for ministry, excluding archived application when a report generation request is made with the appropriate filters.", async () => {
-    // Arrange
-    const now = new Date();
-    const student = await saveFakeStudent(db.dataSource);
-    const institution = await db.institution.save(createFakeInstitution());
-    const [application] = await Promise.all([
-      createApplicationsByInstitutionDataSetup(db, {
-        student,
-        institution,
-        originalSubmissionDate: now,
-      }),
-      createApplicationsByInstitutionDataSetup(db, {
-        student,
-        institution,
-        originalSubmissionDate: now,
-        // Set the offering end date to be outside the archive date range to ensure that the application is archived.
-        currentOfferingEndDateOffSet: -applicationArchiveDays - 1,
-      }),
-    ]);
-    const payload = {
-      reportName: "Ministry_Student_Applications_By_Institution_Report",
-      params: {
-        institution: institution.id,
-        program: "",
-        startDate: now,
-        endDate: addDays(1, now),
-        isLimitedByArchiveDate: true,
-        offeringIntensity: {
-          "Full Time": true,
-          "Part Time": true,
-        },
-      },
-    };
-
-    const endpoint = "/aest/report";
-    const ministryUserToken = await getAESTToken(
-      AESTGroups.BusinessAdministrators,
-    );
-    // Expected report records.
-    const expectedRecord = buildApplicationsByInstitutionData(application);
-
-    // Act/Assert
-    await request(app.getHttpServer())
-      .post(endpoint)
-      .send(payload)
-      .auth(ministryUserToken, BEARER_AUTH_TYPE)
-      .expect(HttpStatus.CREATED)
-      .then((response) => {
-        const fileContent = response.request.res["text"];
-        const parsedResult = parse(fileContent, {
-          header: true,
-        });
-        expect(parsedResult.data).toEqual([expectedRecord]);
-      });
-  });
 
   /**
    * Converts education program offering object into a key-value pair object matching the result data.
